move db stuff into own file to export, sign up page inserts user to db, broke front end messages
