+++ app.js
@@ -1,44 +1,24 @@
 // Setup
-const express = require("express");
-const path = require("path");
-const mysql = require("mysql");
-const dotenv = require("dotenv");
+const express = require('express');
+const path = require('path');
+var db = require('./db');
 
 const app = express();
-
-dotenv.config({ path: './.env'});
-
-// keep settings in local so everyone can setup their stuff indpendently
-const db = mysql.createConnection({
-    host: process.env.DATABASE_HOST,
-    port: process.env.DATABASE_PORT,
-    database: process.env.DATABASE,
-    user: process.env.DATABASE_USER,
-    password: process.env.DATABASE_PASSWORD
-});
 
 const publicDirectory = path.join(__dirname, './public');
 app.use(express.static(publicDirectory));
 
 // Parse URL-encoded bodies (as sent by forms)
-app.use(express.urlencoded({ extended: false}));
+app.use(express.urlencoded({ extended: false }));
 // Parse JSON bodies
 app.use(express.json());
 
 app.set('view engine', 'hbs');
-
-db.connect( (error) => {
-    if(error){
-        console.log(error);
-    } else {
-        console.log("MySQL connected!");
-    }
-});
 
 // Define Routes
 app.use('/', require('./routes/pages'));
 app.use('/auth', require('./routes/auth'));
 
 app.listen(5000, () => {
-    console.log("Server started on port 5000");
+    console.log('Server started on port 5000');
 });
