+++ package.json
@@ -5,14 +5,17 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start" : "nodemon app.js"
+    "start": "nodemon app.js"
   },
   "dependencies": {
-    "express" : "^4.17.1",
-    "mysql" : "^2.18.1",
-    "dotenv" : "^10.0.0",
-    "hbs" : "^4.1.2",
-    "nodemon" : "^2.0.12"
+    "bcryptjs": "^2.4.3",
+    "cookie-parser": "^1.4.5",
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "hbs": "^4.1.2",
+    "jsonwebtoken": "^8.5.1",
+    "mysql": "^2.18.1",
+    "nodemon": "^2.0.12"
   },
   "keywords": [],
   "author": "Charlie Taylor",
