+++ views/index.hbs
@@ -34,6 +34,11 @@
                 <label for="floatingPassword">Password</label>
             </div>
 
+            <div class="form-floating">
+                <input type="password" class="form-control" id="floatingPasswordConfirm" name="passwordConfirm" placeholder="Confirm Password">
+                <label for="floatingPasswordConfirm">Confirm Password</label>
+            </div>
+
             <div class="checkbox mb-3">
                 <label>
                     <input type="checkbox" value="remember-me"> Remember me
