+++ views/signup.hbs
@@ -37,6 +37,10 @@
         </div>
     </div>
 
+    {{#if message}}
+     <h4 class="alert alert-danger mt-4>{{message}}</h4>
+    {{/if}}
+
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-/bQdsTh/da6pkI1MST/rWKFNjaCP5gBSY4sEBT38Q/9RBh9AH40zEOg7Hlq2THRZ" crossorigin="anonymous"></script>
 </body>
 </html>
